+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - DOM tool: Remove trailling .00 in round DOM dimensions
 - Character tool: Add option to hide the number of characters
+- Minify: Fix wrong space remove between % and -
 
 # 1.1.5